Fixed a bug which mistakenly label finished simulation as STALL
If the user uses ic_generator to generate a parameter space, then uses
SiMon to finish the parameter space. Then generates a new parameter
space in the same directory, the finished simulations will be
mistakenly marked as STALL, because the per-simulation configuration
files are overwritten and the default pid=0, which always exists.